pyproject.toml
[project]
name = "gardena-smart-local-api"
version = "0.0.1"
description = "Library to interact with the local GARDENA smart API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofile>=3.9.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
    "ty>=0.0.16",
]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["gardena_smart_local_api"]

[tool.coverage.run]
relative_files = true
source = ["gardena_smart_local_api"]

[tool.coverage.report]
show_missing = true
